fix: refactor warehouse to implement necesarry changes in employee and inbound_orders
