+++ coverage.out
@@ -1,0 +1,34 @@
+mode: set
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:22.69,24.2 1 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:27.74,29.46 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:29.46,32.3 2 0
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:34.2,39.28 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:39.28,41.3 1 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:41.8,43.3 1 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:45.2,46.16 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:46.16,49.3 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:51.2,52.51 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:56.74,58.16 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:58.16,61.3 2 0
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:62.2,62.25 1 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:62.25,66.3 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:67.2,68.32 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:68.32,70.3 1 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:71.2,71.47 1 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:75.75,78.16 3 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:78.16,81.3 2 0
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:82.2,83.16 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:83.16,86.3 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:87.2,88.46 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:92.74,95.16 3 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:95.16,98.3 2 0
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:99.2,102.43 4 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:102.43,105.3 2 0
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:106.2,107.16 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:107.16,110.3 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:111.2,112.46 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:116.74,119.16 3 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:119.16,122.3 2 0
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:123.2,123.58 1 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:123.58,126.3 2 1
+github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/employee/employee.go:127.2,127.37 1 1
