+++ internal/service/inbound_order/inbound_order.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"errors"
 
-	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/repository"
 	empRepo "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/repository/employee"
 	inbRepo "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/repository/inbound_order"
+	wRepo "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/repository/warehouse"
 	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/validators"
 	"github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/api/apperrors"
 	models "github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/models/inbound_order"
@@ -16,14 +16,14 @@ import (
 type InboundOrderDefault struct {
 	repo          inbRepo.InboundOrderRepository
 	employeeRepo  empRepo.EmployeeRepository
-	warehouseRepo repository.WarehouseRepository
+	warehouseRepo wRepo.WarehouseRepository
 }
 
 // Constructor del servicio de inbound orders
 func NewInboundOrderService(
 	r inbRepo.InboundOrderRepository,
 	er empRepo.EmployeeRepository,
-	wr repository.WarehouseRepository) *InboundOrderDefault {
+	wr wRepo.WarehouseRepository) *InboundOrderDefault {
 	return &InboundOrderDefault{
 		repo:          r,
 		employeeRepo:  er,
